organized the project into separate source files
