+++ src/collision/collision.hpp
@@ -1,0 +1,7 @@
+#include "../entities/entities.hpp"
+
+void playerPlatformCollision(Player *player, Platform *platform,
+                             CollisionState *collisionState);
+
+void playerWallCollision(Player *player, Wall walls[45],
+                          CollisionState *collisionState);
