+++ src/keybindings/keybindings.hpp
@@ -1,0 +1,5 @@
+#include "../entities/entities.hpp"
+
+bool clickKeybindings(SDL_Event event, MotionState *motionState, CollisionState *collisionState);
+
+void holdKeybindings(Player *player, SDL_GameController *gamecontroller);
